test(io): cast slong to long long in flint_sprintf expected values
Caught by the MinGW64 (LLP64) CI run on the previous commit (141266ad23):
slong on Windows is `long long` (64-bit) but `long` is 32-bit, so the
expected-value computation `snprintf(expected, ..., "%ld", (long) values[ix])`
truncated WORD_MIN/WORD_MAX to 0 and the test reported a false failure
"flint_sprintf(\"[%wd]\", -9223372036854775808) gave ... expected \"[0]\"".

Cast to `long long` and use `%lld` instead, which fits slong on every
supported platform (slong is `long` on LP64 and `long long` on LLP64).

+++ src/test/t-io.c
@@ -1450,7 +1450,7 @@ TEST_FUNCTION_START(flint_sprintf, state)
         char expected[16];
         memset(buf, 'Z', sizeof(buf));
         got = flint_sprintf(buf, "x%wd", i + 1);
-        snprintf(expected, sizeof(expected), "x%ld", (long) (i + 1));
+        snprintf(expected, sizeof(expected), "x%lld", (long long) (i + 1));
         if (got < 0)
             TEST_FUNCTION_FAIL(
                     "Negative return value from flint_sprintf for i=%wd.\n", i);
@@ -1468,7 +1468,7 @@ TEST_FUNCTION_START(flint_sprintf, state)
             char expected[64];
             memset(buf, 'Z', sizeof(buf));
             got = flint_sprintf(buf, "[%wd]", values[ix]);
-            snprintf(expected, sizeof(expected), "[%ld]", (long) values[ix]);
+            snprintf(expected, sizeof(expected), "[%lld]", (long long) values[ix]);
             if (got < 0 || strcmp(buf, expected) != 0)
                 TEST_FUNCTION_FAIL(
                         "flint_sprintf(\"[%%wd]\", %wd) gave \"%s\" expected \"%s\"\n",
